+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+* 0.1.3:
+  * **Fix** smouth scroll
+  * **Fix** zoom to target top limit
 * 0.1.2:
   * **Fix** world map anchor
 * 0.1.1: